[Feat/#16] : 곱셈 매개변수 자료형 수정

+++ calculator/src/Multiplication.java
@@ -1,6 +1,6 @@
 //곱셈 기능 담당 클래스
 public class Multiplication {
-    public static double multiply(int num1, int num2) {    //실수형 매개변수 2개 받는 매서드 static으로 선언
+    public static double multiply(double num1, double num2) {    //실수형 매개변수 2개 받는 매서드 static으로 선언
         return num1 * num2;                                //  num1과 num2를 곱한 값 반환
     }
 }
